[feat]:recycle: hooks 파일 리팩토링
